+++ src/main/java/roomescape/controller/dto/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record CreateReservationRequest(String date, String name, Long timeId) {
