+++ src/main/java/roomescape/controller/dto/CreateReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record CreateReservationTimeRequest(String startAt) {
