+++ src/main/java/roomescape/controller/dto/CreateReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record CreateReservationTimeResponse(Long id, String startAt) {
