+++ src/main/java/roomescape/controller/dto/FindReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record FindReservationTimeResponse(Long id, String startAt) {
