+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.dto.CreateReservationTimeRequest;
+import roomescape.controller.dto.CreateReservationTimeRequest;
 import roomescape.domain.ReservationTime;
 
 @Repository
