+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,7 +2,7 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.dto.CreateReservationTimeRequest;
+import roomescape.controller.dto.CreateReservationTimeRequest;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationTimeRepository;
